Refactor: Move code out of __main__.py
Following @dalonsoa's suggestions.
